example app tweak - make the code work for both debug and release builds, while including dbinspector in debug only. The change is also necessary for "gradlew check" to pass.
