+++ ExampleApp/src/release/java/im/dino/dbview/exampleapp/ExampleTablePageFormatterFactory.java
@@ -1,0 +1,10 @@
+package im.dino.dbview.exampleapp;
+
+class ExampleTablePageFormatterFactory {
+    public static void init() {
+        // No-op in release, see the version in debug source tree for actual implementation.
+        //
+        // This entry point method MUST be devoid of any references to dbinspector, so that
+        // dbinspector can be included as as debug-only dependency.
+    }
+}
